feat: Add YamlReader
This class converts yaml files to java objects with snakeyaml
