--- create_jar_file.sh
@@ -1,3 +1,0 @@
-javac -d target/classes src/main/java/com/analyzer/App.java
-
-jar uf target/analyzer-1.0-SNAPSHOT.jar -C target/classes .
